+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Insurance feature.
 
+### Modified
+- Custom add card delay for the dealer.
+
 ### Changed
 - Layout.